change base font for mobile
Changing the font-size on the html element to 22px turns 0.75rem to roughly 16px (16px = 0.7273rem, so 0.75 is like 16.5-ish px), which then prevents the iOS browser from zooming into the input fields.

This change obviously makes the whole website a bit bigger and breaks things on Desktop, so we change it back to 18px with the first breakpoint.