Add nginx container to resolve CORS errors (#432)
* Added #! for oneoff scripts.

This should allow us to run them without using `python -m`.

Also reorganized a couple imports so system imports are first: https://peps.python.org/pep-0008/#imports

* Update setub_db.sh

Allowing this to fail individual steps without crashing the whole thing.

The database setup will fail if there's already stuff in it. That might be an improvement we can make later.

* Fixed some pathes in our tiltfile

We were doing these updates semi-correctly before, but now it'll behave more properly - we were doing a bit more building than we needed to.

* Added .gitattributes

When checking out dofuslab on a Windows machine, we can have issues with copying CRLF-line-ending files to Linux containers. This should fix that.

* Added a bunch of ::words:: about using Tilt/Docker

Update to the readme file. This will be simplified further when we get better live-updates working for database updates. We might need more ::words:: to help support would-be contributors, though.

* Changed server home page to docker internal address.

This should work to remove the requirement to add dev.localhost to /etc/hosts.

* Adding reset-db scripts.

It should now be easy to reset the database if you bork something.

It'll just take a min. Or three.

* (WIP) Some notes on live syncing database entries.

Adding some commented syncs for database stuff. Gonna have to add parameters to those sync scripts to make them do everything in those tables.

* Fixing permissions on .sh scripts.

* Fixed init script mount/init of database.

* Update package.json to invoke apollo via yarn.

* Fixed build cache for NODE_ENV in client container

Also defaulting the client container to `development` rather than production.

If you run the container as production it won't find `apollo` and crash.

* Fixing line endings.

* Resolve CORS issues with docker

Added a new nginx docker container that acts as a reverse proxy, so we don't have to disable CORS warnings for development.

Also adjusted our environment variables and README to match the changes.

* Update Tiltfile service links, nginx rebuild